configure `gpg-agent` and use it for ssh 🕵️
In addition to providing a `~/.gnupg/gpg-agent.conf` file and starting
the `gpg-agent` when a login shell opens, this adds configuration for
`gpg-agent` to replace `ssh-agent`. This replacement is needed in order
to have a YubiKey-managed RSA key used while setting up ssh sessions.

Additionally, set up some loose management of ~/.ssh/config to reflect
this - with a fallback option if a non-YubiKey-managed key named
`id_ed25519_github` exists instead.

In the future, much of this may be able to be removed: ~/.ssh/config
should be managed directly (outside of the context of gpg) _and_ the
`gpg-agent` should not be needed in the loop for ssh if an
`ed25519-sk`-type key is used
(https://docs.github.com/en/authentication/connecting-to-github-with-ssh/generating-a-new-ssh-key-and-adding-it-to-the-ssh-agent#generating-a-new-ssh-key-for-a-hardware-security-key).