do not alert on missing identity when using sk 📵
Potential `IdentityFile`s from an entry in `~/.ssh/config` are tried
sequentially, and if a key file is missing a `no such identity` output is
printed.

During ssh key setup: maintain separate names for `-sk` vs. non-`-sk`
keys, but symlink the non-`-sk` name to the `-sk` name if a
YubiKey-backed key is created.

Relying on the symlink, simplify `~/.ssh/config` such that only one
`IdentityFile` is checked for GitHub, eliminating the error when a key
is missing.